+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v1.0.4 - TBD
 
+# Fixed
+
+- [#110](https://github.com/hyperf-cloud/hyperf/pull/110) Fixed Redis::select is not work expected.
 
 # v1.0.3 - 2019-07-02